+++ src/matterbridge.ts
@@ -45,7 +45,7 @@ import { Frontend } from './frontend.js';
 
 // @matter
 import { DeviceTypeId, Endpoint as EndpointNode, Logger, LogLevel as MatterLogLevel, LogFormat as MatterLogFormat, VendorId, StorageContext, StorageManager, StorageService, Environment, ServerNode, FabricIndex, SessionsBehavior } from '@matter/main';
-import { ExposedFabricInformation, FabricAction, MdnsService, PaseClient } from '@matter/main/protocol';
+import { DeviceCommissioner, ExposedFabricInformation, FabricAction, MdnsService, PaseClient } from '@matter/main/protocol';
 import { AggregatorEndpoint } from '@matter/main/endpoints';
 
 // Default colors
@@ -2157,6 +2157,16 @@ export class Matterbridge extends EventEmitter {
     this.log.notice(`Closing ${matterServerNode.id} server node`);
     await matterServerNode.close();
     await matterServerNode.env.get(MdnsService)[Symbol.asyncDispose]();
+  }
+
+  async advertiseServerNode(matterServerNode?: ServerNode) {
+    if (matterServerNode && matterServerNode.lifecycle.isCommissioned) {
+      await matterServerNode.env.get(DeviceCommissioner)?.allowBasicCommissioning();
+      const { qrPairingCode, manualPairingCode } = matterServerNode.state.commissioning.pairingCodes;
+      this.log.notice(`Advertising for ${matterServerNode.id} is now started with the following pairing codes: qrPairingCode ${qrPairingCode}, manualPairingCode ${manualPairingCode}`);
+      return { qrPairingCode, manualPairingCode };
+    }
+    return undefined;
   }
 
   private async createAggregatorNode(storageContext: StorageContext) {